DataManager.jl API

EngineeringDataManager.DataManager.fill_ids — Method

fill_ids()

Fills ID => name dictionary IDS[] and NAMES[] dictionary for name => ID Name can contain several ids thus NAMES[] dictionary elements are the vectors of strings

EngineeringDataManager.DataManager.fill_materials_nodes — Method

fill_materials_nodes()

Fills materials nodes dictionary

EngineeringDataManager.DataManager.get_all_qualifiers — Method

get_all_qualifiers(w_node::Union{ParameterValueNode,PropertyDataNode})

Gets all qualifiers from the node and puts them to the OrderedDict with qualifier name as a key and qualifier string as a value.

EngineeringDataManager.DataManager.get_material_node — Method

get_material_node(material_name)

Material node by name, returns Tuple (materialname,materialnode)

EngineeringDataManager.DataManager.get_node_data — Method

get_node_data(w_node::ParameterValueNode)

Returns nodes data field content for a single node of ParameterValueNode type

Returns named tuple: (name = param_name, - the name of the parameter id = id, - paramater/property id data = data, - data (parsed to vector) qualifier = qual - qualifier )

EngineeringDataManager.DataManager.get_prop_id_by_prop_name — Method

get_prop_id_by_prop_name(property_name)

As far as NAMES dictionary stores both properties and parameters associated to the property name, this function is used to get property id pr* from the property name (otherwise it returns error)

EngineeringDataManager.DataManager.get_property_node — Method

get_property_node(property_id_or_name,material_name::Union{String,Nothing}=nothing)

Returns node of property by ID or by its name, for material material_name or for all materials if nothing

EngineeringDataManager.DataManager.is_property_data_node — Method

is_property_data(node)

Checks if the node can be wrapped to PropertyData data type

EngineeringDataManager.DataManager.material_nodes — Method

material_nodes(;material_name::String = "")

Returns a tuple of two vectors, first - materials names, second one - nodes for material material_name, if material_name = "" returns nodes for all materials

EngineeringDataManager.DataManager.search_by_id — Method

search_by_id(starting_node,id::AbstractVector)

Returns property nodes vector by IDs of parameters

EngineeringDataManager.DataManager.search_by_id — Method

search_by_id(starting_node,id::String)

Returns nodes vector by parameter ID