DataServer.jl API

EngineeringDataManager.DataServer — Module

Module which runs a simple tcp/ip server to elaborate requests from external clients

Requests are configured in [COMMANDS_LIST] dictionary. First, client should send a string of request (key of the COMMANDS_LIST), further is should conretize the request according to the request callback specification.

EngineeringDataManager.DataServer.request_port_names — Method

request_port_names(serv::TCP_Server,sock::TCPSocket)

Returns all clients ports names

EngineeringDataManager.DataServer.request_property_data — Method

request_property_data(::TCP_Server,socket::TCPSocket)

Client property request callback

EngineeringDataManager.DataServer.start_server — Method

start_server(port)

Starts localhost server on specified port

EngineeringDataManager.DataServer.start_server — Method

start_server(;port = DEFAULT_PORT)

Starts localhost server on specified port

EngineeringDataManager.DataServer.stop_server — Method

stop_server(serv::TCP_Server,::TCPSocket)

Server shutting down